refactor: uglify && terser 配置抽离
