+++ lerna.json
@@ -12,7 +12,8 @@
     "packages/taro-plugin-babel",
     "packages/taro-plugin-csso",
     "packages/taro-plugin-sass",
-    "packages/taro-plugin-uglifyjs",
+    "packages/taro-plugin-terser",
+    "packages/taro-plugin-uglify",
     "packages/taro-plugin-typescript",
     "packages/taro-plugin-less",
     "packages/taro-plugin-stylus",
