+++ packages/taro-plugin-terser/index.js
@@ -1,0 +1,2 @@
+module.exports = require('./dist/index.js').default
+module.exports.default = module.exports
