+++ packages/taro-plugin-terser/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "@tarojs/plugin-terser",
+  "version": "2.2.7",
+  "description": "Taro webpack terser config plugin",
+  "main": "index.js",
+  "scripts": {
+    "build": "npm run clean && npm run prod",
+    "dev": "tsc -w",
+    "prod": "tsc",
+    "clean": "rimraf dist",
+    "test": "jest",
+    "test:dev": "jest --watch"
+  },
+  "files": [
+    "src",
+    "dist",
+    "index.js",
+    "package.json"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/NervJS/taro.git"
+  },
+  "keywords": [
+    "taro"
+  ],
+  "author": "luckyadam",
+  "license": "MIT",
+  "dependencies": {
+    "@tarojs/runner-utils": "2.2.7",
+    "@tarojs/service": "2.2.7",
+    "terser-webpack-plugin": "3.0.2"
+  },
+  "devDependencies": {
+    "@types/node": "^14.0.5",
+    "jest": "^26.0.1",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^26.0.0",
+    "typescript": "^3.9.3"
+  }
+}
