+++ packages/taro-plugin-uglify/index.js
@@ -1,0 +1,2 @@
+module.exports = require('./dist/index.js').default
+module.exports.default = module.exports
