--- packages/taro-plugin-uglifyjs/index.js
@@ -1,5 +1,0 @@
-const uglifyJS = require('uglify-js')
-
-module.exports = function uglify (content, file, config) {
-  return uglifyJS.minify(content, config)
-}
