--- packages/taro-plugin-uglifyjs/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@tarojs/plugin-uglifyjs",
-  "version": "2.2.7",
-  "description": "Taro压缩JS文件",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/NervJS/taro.git"
-  },
-  "keywords": [
-    "taro"
-  ],
-  "author": "luckyadam",
-  "license": "MIT",
-  "dependencies": {
-    "uglify-js": "^3.3.24"
-  }
-}
